+++ src/components/BarChart/stories/utils.stories.tsx
@@ -13,7 +13,10 @@ export const colorOptions: string[] = Object.keys(vizColors).concat(
 );
 
 export const getDataPoint = (limit = 1000, allowNegative = false) => {
-  if (allowNegative) return Math.ceil(Math.random() * limit) * (Math.round(Math.random()) ? 1 : -1)
+  if (allowNegative)
+    return (
+      Math.ceil(Math.random() * limit) * (Math.round(Math.random()) ? 1 : -1)
+    );
   return Math.random() * limit;
 };